# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rail Slot Planning MVP API
  version: 1.0.0
  description: All operational values in canonical-v1 are DEMO_ASSUMPTION.
paths:
  /v1/scenarios:
    post:
      operationId: createScenario
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ScenarioCreateRequest'}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Scenario'}}}}
        '400': {$ref: '#/components/responses/InvalidInput'}
  /v1/scenarios/{scenario_id}/validate:
    post:
      operationId: validateScenario
      parameters: [{$ref: '#/components/parameters/ScenarioId'}]
      responses:
        '200': {description: Validation completed, content: {application/json: {schema: {$ref: '#/components/schemas/ValidationResult'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/scenarios/{scenario_id}/runs:
    post:
      operationId: runBaselinePlan
      parameters: [{$ref: '#/components/parameters/ScenarioId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/RunRequest'}}}
      responses:
        '201': {description: Run completed, content: {application/json: {schema: {$ref: '#/components/schemas/Run'}}}}
        '400': {$ref: '#/components/responses/InvalidInput'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationRequired'}
        '503': {$ref: '#/components/responses/SolverUnavailable'}
  /v1/runs/{run_id}:
    get:
      operationId: getRun
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Run result, content: {application/json: {schema: {$ref: '#/components/schemas/Run'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/runs/{run_id}/alternatives:
    post:
      operationId: createAlternative
      parameters: [{$ref: '#/components/parameters/RunId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/AlternativeRequest'}}}
      responses:
        '201': {description: Alternative run, content: {application/json: {schema: {$ref: '#/components/schemas/AlternativeResult'}}}}
        '200': {description: Permitted alternatives were searched but none is feasible, content: {application/json: {schema: {$ref: '#/components/schemas/AlternativeUnavailableResult'}}}}
        '400': {$ref: '#/components/responses/InvalidInput'}
        '409': {$ref: '#/components/responses/PolicyViolation'}
  /v1/runs/{run_id}/decisions:
    post:
      operationId: recordDecision
      parameters: [{$ref: '#/components/parameters/RunId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/DecisionRequest'}}}
      responses:
        '201': {description: Decision recorded, content: {application/json: {schema: {$ref: '#/components/schemas/Decision'}}}}
        '409': {$ref: '#/components/responses/RunNotAcceptable'}
  /v1/runs/{run_id}/export:
    get:
      operationId: exportRunBundle
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Immutable demo and verification bundle, content: {application/json: {schema: {$ref: '#/components/schemas/ExportBundle'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationRequired'}
...
components:
  parameters:
    ScenarioId: {name: scenario_id, in: path, required: true, schema: {type: string}}
    RunId: {name: run_id, in: path, required: true, schema: {type: string}}
  responses:
    InvalidInput: {description: Invalid input, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    NotFound: {description: Resource not found, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    ValidationRequired: {description: Scenario must be validated first, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    PolicyViolation: {description: Requested change is forbidden, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    RunNotAcceptable: {description: Only OPTIMAL and validator PASS runs can be accepted, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    SolverUnavailable: {description: The solver produced no plan, so no run was created, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  schemas:
    ScenarioCreateRequest:
      type: object
      required: [scenario_name, as_of, baseline_service_ids, input_snapshot, policy_version]
      properties:
        scenario_name: {type: string}
        as_of: {type: string, format: date-time}
        baseline_service_ids: {type: array, items: {type: string}, minItems: 1}
        input_snapshot: {$ref: '#/components/schemas/ScenarioInputSnapshot'}
        policy_version: {type: string}
        assumption_ids: {type: array, items: {type: string}}
    Scenario:
      type: object
      required: [scenario_id, state, created_at]
      properties:
        scenario_id: {type: string}
        state: {type: string, enum: [VALIDATION_REQUIRED, READY_TO_SOLVE, SOLVED]}
        created_at: {type: string, format: date-time}
    ScenarioInputSnapshot:
      type: object
      required: [schema_version, scenario_id, scenario_type, as_of, timezone, baseline_service_ids, assumptions, shippers, terminals, route_constraints, services, wagons, slots, orders, policy]
      properties:
        schema_version: {type: string, const: '1.0.0'}
        scenario_id: {type: string, minLength: 1}
        scenario_type: {type: string, enum: [BASELINE, ALTERNATIVE]}
        as_of: {type: string, format: date-time}
        timezone: {type: string, const: Asia/Seoul}
        baseline_service_ids: {type: array, minItems: 1, items: {type: string}}
        assumptions: {type: array, minItems: 1, items: {$ref: '#/components/schemas/Assumption'}}
        shippers: {type: array, minItems: 1, items: {$ref: '#/components/schemas/Shipper'}}
        terminals: {type: array, minItems: 1, items: {$ref: '#/components/schemas/Terminal'}}
        route_constraints: {type: array, minItems: 1, items: {$ref: '#/components/schemas/RouteConstraint'}}
        services: {type: array, minItems: 1, items: {$ref: '#/components/schemas/Service'}}
        wagons: {type: array, minItems: 1, items: {$ref: '#/components/schemas/Wagon'}}
        slots: {type: array, minItems: 1, items: {$ref: '#/components/schemas/Slot'}}
        orders: {type: array, minItems: 1, items: {$ref: '#/components/schemas/Order'}}
        policy: {$ref: '#/components/schemas/Policy'}
...
    SolverParameters:
      type: object
      required: [random_seed, num_search_workers, max_time_seconds]
      properties:
        random_seed: {type: integer, minimum: 0}
        num_search_workers: {type: integer, const: 1}
        max_time_seconds: {type: integer, minimum: 1}
    RunRequest:
      type: object
      required: [solver_parameters]
      properties:
        solver_parameters: {$ref: '#/components/schemas/SolverParameters'}
    ValidationResult:
      type: object
      required: [scenario_id, validation_status, orders]
      properties:
        scenario_id: {type: string}
        validation_status: {type: string, enum: [COMPLETED, FAILED]}
        orders: {type: array, items: {$ref: '#/components/schemas/OrderValidation'}}
    OrderValidation:
      type: object
      required: [order_id, input_state, reason_codes, eligible_slot_ids, eligibility_state]
      properties:
        order_id: {type: string}
        input_state: {type: string, enum: [VALID, REVIEW_REQUIRED]}
        eligibility_state: {type: string, enum: [ELIGIBLE, INELIGIBLE, NOT_EVALUATED]}
        reason_codes: {type: array, items: {type: string}}
        missing_fields: {type: array, items: {type: string}}
        eligible_slot_ids: {type: array, items: {type: string}}
        primary_reason_code:
          anyOf: [{type: string}, {type: 'null'}]
          description: The single reason to show, chosen by the 02 §6.1 family priority.
    Run:
      type: object
      required: [run_id, scenario_id, solver_status, run_state, is_optimal, validator_status, reproducibility, assignments, order_outcomes]
      properties:
        run_id: {type: string}
        scenario_id: {type: string}
        # ERROR is deliberately absent: a solve that proves nothing yields
        # 503 SOLVER_UNAVAILABLE and no run, so no client can ever observe it.
        solver_status: {type: string, enum: [OPTIMAL, FEASIBLE, INFEASIBLE]}
        run_state: {type: string, enum: [SOLVED_OPTIMAL, SOLVED_FEASIBLE, MODEL_INFEASIBLE]}
        is_optimal: {type: boolean}
        validator_status: {type: string, enum: [PASS, FAIL]}
        validator_findings:
          type: array
          items: {$ref: '#/components/schemas/ValidatorFinding'}
          description: >-
            Empty on PASS. On FAIL the UI must name the broken rule and the
            resource it broke on (08 §9), which a bare status cannot do.
        reproducibility: {$ref: '#/components/schemas/Reproducibility'}
        assignments: {type: array, items: {$ref: '#/components/schemas/Assignment'}}
        order_outcomes: {type: array, items: {$ref: '#/components/schemas/OrderOutcome'}}
        objective_values:
          type: object
          additionalProperties: {type: integer}
          description: >-
            Value reached at each lexicographic stage. The only signal that
            separates a fully refined OPTIMAL plan from one whose time budget
            ran out mid-refinement.
    ValidatorFinding:
      type: object
      required: [check, order_id, message]
      properties:
        check: {type: string}
        order_id: {anyOf: [{type: string}, {type: 'null'}]}
        message: {type: string}
    Reproducibility:
      type: object
      required: [solver_parameters, input_snapshot_sha256, policy_sha256, result_sha256]
      properties:
        solver_parameters: {$ref: '#/components/schemas/SolverParameters'}
        input_snapshot_sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
        policy_sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
        result_sha256: {type: string, pattern: '^[a-f0-9]{64}$'}
    Assignment:
      type: object
      required: [order_id, service_id, wagon_id, slot_id]
      properties:
        order_id: {type: string}
        service_id: {type: string}
        wagon_id: {type: string}
        slot_id: {type: string}
    AssignmentDelta:
      type: object
      required: [order_id, change_type, before_assignment, after_assignment]
      properties:
        order_id: {type: string}
        change_type: {type: string, enum: [ADDED, MOVED, UNASSIGNED]}
        before_assignment: {anyOf: [{$ref: '#/components/schemas/Assignment'}, {type: 'null'}]}
        after_assignment: {anyOf: [{$ref: '#/components/schemas/Assignment'}, {type: 'null'}]}
    OrderOutcome:
      type: object
      required: [order_id, input_state, eligibility_state, assignment_state, alternative_state, primary_reason_code]
      properties:
        order_id: {type: string}
        input_state: {type: string, enum: [VALID, REVIEW_REQUIRED]}
        eligibility_state: {type: string, enum: [ELIGIBLE, INELIGIBLE, NOT_EVALUATED]}
        assignment_state: {type: string, enum: [ASSIGNED, UNASSIGNED, NOT_APPLICABLE]}
        alternative_state: {type: string, enum: [AVAILABLE, NONE, NOT_SEARCHED]}
        primary_reason_code: {type: string}
        evidence: {type: object}
        next_actions: {type: array, items: {type: string}}
        display_label:
          anyOf: [{type: string}, {type: 'null'}]
          description: Computed from the state axes by 02 §4. Never generated.
        display_badges:
          type: array
          items: {type: string}
          description: Additive badges. `조건부 대안 있음` when alternative_state is AVAILABLE.
        alternative_scenario_id:
          anyOf: [{type: string}, {type: 'null'}]
          description: >-
            Set on a baseline order once an alternative is found; the derived
            scenario that can carry it. This is the resource path the
            alternative-comparison screen navigates to.
    AlternativeRequest:
      type: object
      required: [order_id, adjustment_types]
      properties:
        order_id: {type: string}
        adjustment_types: {type: array, minItems: 1, items: {type: string, enum: [ADD_ORDER_APPROVED_SERVICE, CHANGE_TO_APPROVED_TERMINAL, CHANGE_WEIGHT_LIMIT, CHANGE_DIMENSION_LIMIT, CHANGE_ROUTE_CLEARANCE, CHANGE_DUE_AT]}}
    AlternativeResult:
      type: object
      required: [parent_run_id, alternative_scenario_id, alternative_run_id, change_set, impacted_order_ids, baseline_order_update, alternative_run_order_outcome, assignment_deltas, validator_status]
      properties:
        parent_run_id: {type: string}
        alternative_scenario_id: {type: string}
        alternative_run_id: {type: string}
        change_set: {type: array, items: {type: object}}
        impacted_order_ids: {type: array, minItems: 1, items: {type: string}}
        baseline_order_update: {$ref: '#/components/schemas/OrderOutcome'}
        alternative_run_order_outcome: {$ref: '#/components/schemas/OrderOutcome'}
        assignment_deltas: {type: array, minItems: 1, items: {$ref: '#/components/schemas/AssignmentDelta'}}
        validator_status: {type: string, enum: [PASS, FAIL]}
    AlternativeUnavailableResult:
      type: object
      required: [order_id, alternative_state, status, change_set, baseline_order_update, reason_code]
      properties:
        order_id: {type: string}
        alternative_state: {type: string, const: NONE}
        status: {type: string, const: NO_FEASIBLE_ALTERNATIVE}
        change_set: {type: array, items: {type: object}}
        baseline_order_update: {$ref: '#/components/schemas/OrderOutcome'}
        reason_code: {type: string}
    DecisionRequest:
      type: object
      required: [decision_state, actor_role, reason, selected_plan]
      properties:
        decision_state: {type: string, enum: [ACCEPTED, HELD, REJECTED]}
        actor_role: {type: string, enum: [SCHEDULING_OPERATOR, PLANNING_OWNER]}
        reason: {type: string, minLength: 1}
        selected_plan: {type: string, enum: [BASELINE, ALTERNATIVE]}
    Decision:
      type: object
      required: [decision_id, run_id, decision_state, created_at]
      properties:
        decision_id: {type: string}
        run_id: {type: string}
        decision_state: {type: string}
        created_at: {type: string, format: date-time}
    ExportBundle:
      type: object
      required: [scenario, input_snapshot, policy, run, validation_result, decisions, trace_events]
      properties:
        scenario: {$ref: '#/components/schemas/Scenario'}
        input_snapshot: {$ref: '#/components/schemas/ScenarioInputSnapshot'}
        policy: {$ref: '#/components/schemas/Policy'}
        run: {$ref: '#/components/schemas/Run'}
        validation_result: {$ref: '#/components/schemas/ValidationResult'}
        decisions: {type: array, items: {$ref: '#/components/schemas/Decision'}}
        trace_events: {type: array, items: {$ref: '#/components/schemas/TraceEvent'}}
    TraceEvent:
      type: object
      required: [event_id, event_type, occurred_at, payload]
      properties:
        event_id: {type: string}
        event_type: {type: string, enum: [SCENARIO_CREATED, VALIDATION_COMPLETED, RUN_COMPLETED, ALTERNATIVE_CREATED, DECISION_RECORDED]}
        occurred_at: {type: string, format: date-time}
        payload: {type: object}
...
    Error:
      type: object
      required: [code, message, trace_id]
      properties:
        code: {type: string, enum: [INVALID_INPUT, SCENARIO_NOT_FOUND, RUN_NOT_FOUND, VALIDATION_REQUIRED, POLICY_VIOLATION, RUN_NOT_ACCEPTABLE, SOLVER_UNAVAILABLE]}
        message: {type: string}
        details: {type: array, items: {type: object}}
        trace_id: {type: string}